refactor: lazy-load focus export engine and isolate polars deps
